% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctionsDatetime.R
\name{addISO8601colon}
\alias{addISO8601colon}
\title{Adds a colon to the time offset component of an ISO8601 string}
\usage{
addISO8601colon(ISOstring)
}
\arguments{
\item{ISOstring}{ISO date in string format that may or may not include a colon [character]}
}
\value{
ISO 8601 format string based on degree of time precision and desired time zone 
(choices are either UTC or another, non-UTC time) [character]
}
\description{
Adds a colon to the time offset component of an ISO8601 string that 
doesn't already have it
}
\examples{
ISO_string_out <- addISO8601colon("\%Y-\%m-\%dT\%H:\%M+0600")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
date_ingest_checker.R, handleTimestamp.R, ISOStringPrecformatter.R, and 
timestamp_output_formatter.R related functions for formatting dates in ISO standard format
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Jamie Collins \email{jamesrco@uw.edu} \cr
Amanda Droghini \email{adroghini@alaska.edu} \cr
}
\keyword{date,}
\keyword{datetime,}
\keyword{time,}
\keyword{timestamp}
