% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ingest_checker.R
\name{date_ingest_checker}
\alias{date_ingest_checker}
\title{Checks timestamp inputs and converts to POSIXct R object}
\usage{
date_ingest_checker(x, Format = NULL, precision = NULL, TZ = NULL)
}
\arguments{
\item{x}{some input string that contains the dateTime data [character]}

\item{Format}{a format string we can pass to strptime, or NULL; if NULL, we will use 
the tryformats functionality in as.POSIXct; passed to date_ingest_checker() [character]}

\item{precision}{optional argument that can have values of "day", "minute", "second", 
or "milliSecond" for the input and output; if not specified [character]}

\item{TZ}{optional argument specifying the timezone in which the dateTime data was 
collected;must be in the OlsonNames list; if not specified, uses whatever is given 
in the format string for \%z or \%Z [character]}
}
\value{
a timestamp as POSIXct format [POSIXct]
}
\description{
Function to read in a timestamp and format as POSICct in R to help avoid 
timestamp problems with data QAQC, e.g. daylight savings time changes
}
\examples{
date_asPOSIXct <- date_ingest_checker("2018-01-01T13:30","\%Y-\%m-\%dT\%H:\%M","minute","MDT")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
addISO8601colon.R, handleTimestamp.R, ISOStringPrecformatter.R, and 
timestamp_output_formatter.R related functions for formatting dates in ISO standard format
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Jamie Collins \email{jamesrco@uw.edu} \cr
Amanda Droghini \email{adroghini@alaska.edu} \cr
}
\keyword{date,}
\keyword{datetime,}
\keyword{time,}
\keyword{timestamp}
