% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctionsDatetime.R
\name{handleTimestamp}
\alias{handleTimestamp}
\title{Wrapper function for date_ingest_checker() and timestamp_output_formatter()}
\usage{
handleTimestamp(x, Format = NULL, precision = NULL, TZ = NULL,
  suppOutput = NULL)
}
\arguments{
\item{x}{some input string we believe to be a time stamp; passed to 
date_ingest_checker() [character]}

\item{Format}{a format string we can pass to strptime, or NULL; if NULL, we will use 
the tryformats functionality in as.POSIXct; passed to date_ingest_checker() [character]}

\item{precision}{optional argument passed to date_ingest_checker() and then 
timestamp_output_formatter() [character]}

\item{TZ}{optional argument passed to date_ingest_checker(); otherwise time zone will be 
assessed from the data itself; if a value is supplied for TZ, this will be checked against 
any time zone information present in the data to harmonize [character]}

\item{suppOutput}{("supplemental output"): optional argument passed to request timestamp 
output in "local" or other time zone/locality in addition to UTC; must be something in 
the Olson list [character]}
}
\value{
ISO 8601 format string based on degree of time precision and desired time zone 
(choices are either UTC or another, non-UTC time) [character]
}
\description{
Wrapper function for date_ingest_checker() and timestamp_output_formatter()
}
\examples{
precISODateTime <- handleTimestamp("2018-01-01T13:30","\%Y-\%m-\%dT\%H:\%M","minute","MDT","UTC")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
date_ingest_checker.R, addISO8601colon.R, ISOStringPrecformatter.R, and 
timestamp_output_formatter.R related functions for formatting dates in ISO standard format
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Jamie Collins \email{jamesrco@uw.edu} \cr
Amanda Droghini \email{adroghini@alaska.edu} \cr
}
\keyword{date,}
\keyword{datetime,}
\keyword{time,}
\keyword{timestamp}
