% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctionsDatetime.R
\name{ISOStringPrecformatter}
\alias{ISOStringPrecformatter}
\title{Creates a format string for a desired ISO format precision}
\usage{
ISOStringPrecformatter(precision = NULL, desiredZone = c("UTC",
  "other"))
}
\arguments{
\item{precision}{argument that can have values of "day", "minute", "second", 
or "milliSecond" [character]}

\item{desiredZone}{Desired output timezone [character]}
}
\value{
ISO 8601 format string based on degree of time precision and desired time zone 
(choices are either UTC or another, non-UTC time) [character]
}
\description{
Creates complete ISO 8601 format string based on degree of time precision and 
desired time zone (choices are either UTC or another, non-UTC time); if no precision is 
supplied, generates format string with seconds as default
}
\examples{
formatOutISO_complete <- ISOStringPrecformatter("minute","UTC")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
date_ingest_checker.R, addISO8601colon.R, handleTimestamp.R, and 
timestamp_output_formatter.R related functions for formatting dates in ISO standard format
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Jamie Collins \email{jamesrco@uw.edu} \cr
Amanda Droghini \email{adroghini@alaska.edu} \cr
}
\keyword{date,}
\keyword{datetime,}
\keyword{time,}
\keyword{timestamp}
