% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_output_formatter.R
\name{timestamp_output_formatter}
\alias{timestamp_output_formatter}
\title{Function for making timestamps "pretty" for output}
\usage{
timestamp_output_formatter(x, precision = NULL, suppOutput = NULL)
}
\arguments{
\item{x}{a POSIXct object; in our concept, this wouldideally be the output from 
date_ingest_checker(); however, the function is designed to operate by itself with 
user-supplied arguments [POSIXct]}

\item{precision}{optional argument, may be passed along from the value usersupplied for 
the precision argument in date_ingest_checker() or the wrapper function (once we build it); 
if not supplied, we will report to a precision of seconds 
(consistent with ISO 8601 specification) [character]}

\item{suppOutput}{("supplemental output"): optional argument allowing user to request 
timestamp output in "local" or other time zone/locality in addition to UTC; must be something 
in the Olson list [character]}
}
\value{
ISO 8601 format string based on degree of time precision and desired time zone 
(choices are either UTC or another, non-UTC time) [character]
}
\description{
This function takes a POSIXct object, combines with precision argument 
(if supplied),and returns an ISO 8601 text string that has the appropriate level of 
precision. Optionally, also returns the same datetime for a "alternate" time zone other 
than UTC; this would allow user to easily evaluate whether output makes sense based on 
their knowledge of the input data.
}
\examples{
pretty_date <- timestamp_output_formatter(as.POSIXct("2018-01-01T13:30"),"minute")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
addISO8601colon.R, handleTimestamp.R, ISOStringPrecformatter.R, and 
date_ingest_checker.R related functions for formatting dates in ISO standard format
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Jamie Collins \email{jamesrco@uw.edu} \cr
Amanda Droghini \email{adroghini@alaska.edu} \cr
}
\keyword{date,}
\keyword{datetime,}
\keyword{time,}
\keyword{timestamp}
